requirements.txt
django 1.9
Jinja2
pillow
django-registration